Heatsoak safety updates
These are some safety updates related to the way print cancelation works. Basically heat soak now tracks the the state of the virtual_sd_card. It can tell if you started of canceled a print while heat soak was running and take appropriate action to end the heatsoak. It also does this for the pause/resume state.

The upshot is the script wont misbehave if you don't property set up STOP_HEAT_SOAK or HEAT_SOAK_RESUME as is recommended in the instructions.

This also removes the need to have overrides of CANCEL_PRINT. The overrides cant be chained so the effort is the same for the user if I included the overrides or not as most people have a CANCEL_PRINT override.

The RESUME override is still in as most people don't have one of those. HEAT_SOAK_RESUME wraps up the resume functionality so you can call it easily if you want to override.

Lastly you don't have to put the heatsoak macro last in your macro imports, it can go back to the top.