Fix for virtual_sdcard behaviour
Im not sure what exactly changed but in the most recent version of klipper the sdcard reports that there is no active print when paused. But it has a non-zero file position which is a tell that there is a print active. So this checks if there is a print ongoing by looking at the file position instead of `is_active`